Create App.py
In the bike rental system Python program, the total stock of bikes should be shown. The program should also involve requesting a bike for rent as per the quantity. The exit option should also be there in the program, so the user can exit the program whenever he wants. After making this program and working on this Python project, you will learn the concept of OOPS easily.